[ref] ion: Remove build warnings
Linked Commit: None
ITS: TMBSP-2265

+++ drivers/staging/android/ion/ion_heap.c
@@ -226,7 +226,9 @@ static int ion_heap_deferred_free(void *data)
 {
 	struct ion_heap *heap = data;
 
-pr_info("%s name:%s size:%d\n", __func__, heap->name, heap->free_list_size);
+	pr_info("%s name:%s size:%d\n",
+		__func__, heap->name, (u32)heap->free_list_size);
+
 	while (true) {
 		struct ion_buffer *buffer;
